move InFrontOfTheCanvas
